pre generate first default page
