+++ knowledge/lib/utils.ts
@@ -1,3 +1,4 @@
+import dayjs from "dayjs";
 import slugify from "slugify";
 
 import { SortTypeWindowOption, TimeWindowOption } from "../src/knowledgeTypes";
@@ -87,4 +88,47 @@ export const getDefaultSortedByType = (filtersSelected: { mostRecent: boolean; u
   if (filtersSelected.mostRecent) return SortTypeWindowOption.MOST_RECENT;
   if (filtersSelected.upvotes) return SortTypeWindowOption.UPVOTES_DOWNVOTES;
   return SortTypeWindowOption.NONE;
+};
+
+export const buildSortBy = (upvotes: boolean, mostRecent: boolean) => {
+  if (upvotes) {
+    return "mostHelpful:desc";
+  }
+  if (mostRecent) {
+    return "changedAtMillis:desc";
+  }
+  return "";
+};
+
+export const buildFilterBy = (
+  timeWindow: TimeWindowOption,
+  tags: string,
+  institutions: string,
+  contributors: string,
+  nodeTypes: string,
+  reference: string,
+  label: string
+) => {
+  const filters: string[] = [];
+  let updatedAt: number;
+  if (timeWindow === TimeWindowOption.ThisWeek) {
+    updatedAt = dayjs().subtract(1, "week").valueOf();
+  } else if (timeWindow === TimeWindowOption.ThisMonth) {
+    updatedAt = dayjs().subtract(1, "month").valueOf();
+  } else if (timeWindow === TimeWindowOption.ThisYear) {
+    updatedAt = dayjs().subtract(1, "year").valueOf();
+  } else {
+    updatedAt = dayjs().subtract(10, "year").valueOf();
+  }
+
+  filters.push(`changedAtMillis:>${updatedAt}`);
+
+  if (tags.length > 0) filters.push(`tags: [${tags}]`);
+  if (institutions.length > 0) filters.push(`institutionsNames: [${institutions}]`);
+  if (contributors.length > 0) filters.push(`contributorsNames: [${contributors}]`);
+  if (nodeTypes.length > 0) filters.push(`nodeType: [${nodeTypes}]`);
+  if (reference) filters.push(`titlesReferences: ${reference}`);
+  if (label && label !== "All Sections" && label !== "All Pages") filters.push(`labelsReferences: ${label}`);
+
+  return filters.join("&& ");
 };
