+++ knowledge/package.json
@@ -52,6 +52,7 @@
     "remark": "^14.0.2",
     "remark-html": "^15.0.1",
     "remark-math": "^5.1.1",
+    "sharp": "^0.30.6",
     "slugify": "^1.6.5",
     "typescript": "^4.7.2",
     "typesense": "^1.2.2",
